@@ -1,3 +1,14 @@
+// src/environments/environment.prod.ts
 export const environment = {
-  production: true
+  production: true,
+  firebaseConfig: {
+    apiKey: "AIzaSyBYo_NkQLL5BnImzo04UaGeZZuE67Tgc7I",
+    authDomain: "pozos-realtime.firebaseapp.com",
+    databaseURL: "https://pozos-realtime-default-rtdb.firebasedatabase.app",
+    projectId: "pozos-realtime",
+    storageBucket: "pozos-realtime.firebasestorage.app",
+    messagingSenderId: "956172197839",
+    appId: "1:956172197839:web:30c792aac7a9d643dfcba",
+    measurementId: "G-7PZS7GZ88X"
+  }
 };
